Move top menu bar from editor.js to view-controller.js
+ move css dependencies from editor.css to core.css
+ rename css variables from  --editor-top-bar-XXX to --menu-top-bar-XXX
