+++ src/fontra/views/editor/editor.html
@@ -30,8 +30,8 @@
 
     <div id="global-loader-spinner"></div>
 
+    <div class="top-bar-container cleanable-overlay"></div>
     <div class="editor-container">
-      <div class="top-bar-container cleanable-overlay"></div>
       <div class="sidebar-container cleanable-overlay left">
         <div class="sidebar-resize-gutter" data-grow-direction="right"></div>
       </div>
