+++ src/fontra/views/fontoverview/fontoverview.html
@@ -14,6 +14,7 @@
   <body>
     <modal-dialog></modal-dialog>
     <div id="global-loader-spinner"></div>
+    <div class="top-bar-container cleanable-overlay"></div>
     <div class="main-container">
       <div id="sidebar-container" class="font-overview-sidebar-shadow-box"></div>
       <div id="glyph-cell-view-container"></div>
